refactor(T1): split Diagnostics + Restore-replay out of SFExtendService [BUILD-PASSED]
First T1 decomposition round on the Extend god-object. Two cohesive units
extracted into focused sub-services behind thin forwarders; SFExtendService.cpp
9,515 -> 7,718 lines (-1,797, -19%). Pure/careful moves, no behavior change.
Shipping PackagePlugin compile+cook clean (ExitCode=0). In-game smoke pending.

Slice H - Diagnostics (pure move):
- New USFExtendDiagnosticsService.{h,cpp}: the 4 capture structs + the before/after
  world-snapshot + diff logging (CaptureNearbyBuildables / CapturePreviewSnapshot /
  CapturePostBuildSnapshotAndLogDiff / LogSnapshotDiff). Off any gameplay path.
- Reads the current target + topology via the owning service (GetCurrentTarget /
  GetCurrentTopology); owns PreviewSnapshot + bHasPreviewSnapshot.
- SFExtendService keeps CapturePreviewSnapshot / CapturePostBuildSnapshotAndLogDiff /
  HasPreviewSnapshot forwarders; the SFSubsystem caller is unchanged.

Slice B - Restore-replay (careful move):
- New USFExtendRestoreReplayService.{h,cpp}: Smart Restore Extend clone-topology
  replay (Replay/Tick/OnStateChanged/IsHologramCompatible/ClearSession + the
  Build/Spawn/ClearPreview private helpers + the Kick/Align/Scrub anon helpers).
- This unit reads+writes state still owned by USFExtendService (StoredCloneTopology,
  Json* maps, ScaledExtendClones, HologramService), so the sub-service is a friend and
  operates on that shared state in place via Owner-> rather than migrating it. State
  stays put, so GetLastCloneTopology / IsRestoredCloneTopologyActive remain unchanged
  on the main service.
- The placement helper CalculateRestoredScaledClonePlacement is shared with the
  post-build wiring path (GenerateAndExecuteWiring), so it's promoted to a free
  function declared in the new header.
- SFExtendService keeps forwarders for the 5 externally-called methods; the
  Subsystem + RestoreService call sites are unchanged.
