+++ Source/SmartFoundations/Private/Features/Extend/SFExtendService.cpp
@@ -11,6 +11,8 @@
 #include "Features/Extend/SFExtendTopologyService.h"
 #include "Features/Extend/SFExtendHologramService.h"
 #include "Features/Extend/SFExtendWiringService.h"
+#include "Features/Extend/SFExtendDiagnosticsService.h"
+#include "Features/Extend/SFExtendRestoreReplayService.h"
 #include "Features/Extend/SFExtendCloneTopology.h"
 #include "Features/Extend/SFWiringManifest.h"
 #include "Features/Restore/SFRestoreService.h"
@@ -73,205 +75,6 @@
 #include "EngineUtils.h"  // For TActorIterator
 #include "Components/BoxComponent.h"  // For clearance disabling on child factory holograms
 
-namespace
-{
-    constexpr float SF_RESTORED_PARENT_TRANSFORM_TOLERANCE = 0.1f;
-
-    struct FRestoredScaledClonePlacement
-    {
-        FVector WorldOffset = FVector::ZeroVector;
-        FRotator RotationOffset = FRotator::ZeroRotator;
-    };
-
-    FRestoredScaledClonePlacement CalculateRestoredScaledClonePlacement(
-        const AFGHologram* ParentHologram,
-        const FSFCloneTopology* TemplateTopology,
-        const FSFCounterState& State,
-        int32 GridX,
-        int32 GridY)
-    {
-        FRestoredScaledClonePlacement Placement;
-        if (!ParentHologram)
-        {
-            return Placement;
-        }
-
-        USFBuildableSizeRegistry::Initialize();
-        FVector BuildingSize(800.0f, 800.0f, 400.0f);
-        if (UClass* BuildClass = ParentHologram->GetBuildClass())
-        {
-            BuildingSize = USFBuildableSizeRegistry::GetProfile(BuildClass).DefaultSize;
-        }
-
-        const FRotator ParentRotation = ParentHologram->GetActorRotation();
-        float XDirectionSign = State.GridCounters.X < 0 ? -1.0f : 1.0f;
-
-        if (TemplateTopology)
-        {
-            const FRotator OriginalParentRotation = TemplateTopology->ParentTransform.Rotation.ToFRotator();
-            const FRotator RotationDelta = ParentRotation - OriginalParentRotation;
-            const FVector CapturedStep = RotationDelta.RotateVector(TemplateTopology->WorldOffset.ToFVector());
-            const FVector CapturedLocalStep = ParentRotation.UnrotateVector(CapturedStep);
-            if (!FMath::IsNearlyZero(CapturedLocalStep.X))
-            {
-                XDirectionSign *= FMath::Sign(CapturedLocalStep.X);
-            }
-        }
-
-        const float StepDistance = FMath::Max(1.0f, BuildingSize.X + static_cast<float>(State.SpacingX));
-        FVector LocalOffset = FVector::ZeroVector;
-
-        if (!FMath::IsNearlyZero(State.RotationZ))
-        {
-            const float StepRadians = FMath::Abs(FMath::DegreesToRadians(State.RotationZ));
-            const float Radius = (StepRadians > KINDA_SMALL_NUMBER) ? StepDistance / StepRadians : 0.0f;
-            const float SignRotation = (State.RotationZ >= 0.0f) ? 1.0f : -1.0f;
-
-            auto OffsetAtCloneIndex = [&](int32 CloneIndex) -> FVector
-            {
-                const float AngleDeg = static_cast<float>(CloneIndex) * State.RotationZ;
-                const float AbsAngleRad = FMath::Abs(FMath::DegreesToRadians(AngleDeg));
-
-                FVector Offset;
-                Offset.X = XDirectionSign * Radius * FMath::Sin(AbsAngleRad);
-                Offset.Y = SignRotation * (Radius - Radius * FMath::Cos(AbsAngleRad));
-                Offset.Z = static_cast<float>(State.StepsX * CloneIndex);
-                return Offset;
-            };
-
-            const FVector ParentCloneOffset = OffsetAtCloneIndex(1);
-            const FVector TargetCloneOffset = OffsetAtCloneIndex(GridX + 1);
-            LocalOffset = TargetCloneOffset - ParentCloneOffset;
-            Placement.RotationOffset = FRotator(0.0f, State.RotationZ * XDirectionSign * static_cast<float>(GridX), 0.0f);
-        }
-        else
-        {
-            LocalOffset.X = XDirectionSign * StepDistance * static_cast<float>(GridX);
-            LocalOffset.Z = static_cast<float>(State.StepsX * GridX);
-        }
-
-        if (GridY != 0)
-        {
-            const float YSign = State.GridCounters.Y < 0 ? -1.0f : 1.0f;
-            const float RowDistance = FMath::Max(1.0f, BuildingSize.Y + static_cast<float>(State.SpacingY));
-            LocalOffset.Y += RowDistance * static_cast<float>(GridY) * YSign;
-            LocalOffset.Z += static_cast<float>(State.StepsY * GridY);
-        }
-
-        Placement.WorldOffset = ParentRotation.RotateVector(FVector(LocalOffset.X, LocalOffset.Y, 0.0f));
-        Placement.WorldOffset.Z += LocalOffset.Z;
-        return Placement;
-    }
-
-    uint8 AlignRestoredLaneNormals(FSFCloneHologram& Holo, const FVector& Start, const FVector& End)
-    {
-        if (!Holo.bIsLaneSegment || (Holo.LaneSegmentType != TEXT("belt") && Holo.LaneSegmentType != TEXT("pipe")))
-        {
-            return 0;
-        }
-
-        const FVector LaneDirection = (End - Start).GetSafeNormal();
-        if (LaneDirection.IsNearlyZero())
-        {
-            return 0;
-        }
-
-        uint8 FlippedMask = 0;
-        const FVector StartNormal = Holo.LaneStartNormal.ToFVector();
-        if (!StartNormal.IsNearlyZero() && FVector::DotProduct(StartNormal.GetSafeNormal(), LaneDirection) < -0.01f)
-        {
-            Holo.LaneStartNormal = FSFVec3(-StartNormal);
-            FlippedMask |= 1;
-        }
-
-        const FVector EndNormal = Holo.LaneEndNormal.ToFVector();
-        if (!EndNormal.IsNearlyZero() && FVector::DotProduct(EndNormal.GetSafeNormal(), -LaneDirection) < -0.01f)
-        {
-            Holo.LaneEndNormal = FSFVec3(-EndNormal);
-            FlippedMask |= 2;
-        }
-
-        return FlippedMask;
-    }
-
-    void KickRestoredPreviewParent(AFGHologram* ParentHologram)
-    {
-        if (!IsValid(ParentHologram))
-        {
-            return;
-        }
-
-        const FVector Location = ParentHologram->GetActorLocation();
-        const FRotator Rotation = ParentHologram->GetActorRotation();
-        const bool bWasLocked = ParentHologram->IsHologramLocked();
-
-        FHitResult SyntheticHit;
-        SyntheticHit.bBlockingHit = true;
-        SyntheticHit.Location = Location;
-        SyntheticHit.ImpactPoint = Location;
-        SyntheticHit.ImpactNormal = FVector::UpVector;
-        SyntheticHit.Normal = FVector::UpVector;
-        SyntheticHit.TraceStart = Location + FVector(0.0f, 0.0f, 100.0f);
-        SyntheticHit.TraceEnd = Location - FVector(0.0f, 0.0f, 100.0f);
-        SyntheticHit.Distance = 100.0f;
-        ParentHologram->SetHologramLocationAndRotation(SyntheticHit);
-
-        ParentHologram->SetActorLocation(Location);
-        ParentHologram->SetActorRotation(Rotation);
-        if (USceneComponent* Root = ParentHologram->GetRootComponent())
-        {
-            Root->SetWorldLocation(Location);
-            Root->SetWorldRotation(Rotation);
-            Root->MarkRenderStateDirty();
-        }
-
-        ParentHologram->SetActorHiddenInGame(false);
-        ParentHologram->UpdateComponentTransforms();
-
-        ParentHologram->LockHologramPosition(!bWasLocked);
-        ParentHologram->LockHologramPosition(bWasLocked);
-    }
-
-    void ScrubInvalidHologramChildren(AFGHologram* ParentHologram, const TCHAR* Context)
-    {
-        if (!IsValid(ParentHologram))
-        {
-            return;
-        }
-
-        FArrayProperty* ChildrenProp = FindFProperty<FArrayProperty>(AFGHologram::StaticClass(), TEXT("mChildren"));
-        if (!ChildrenProp)
-        {
-            return;
-        }
-
-        TArray<AFGHologram*>* ChildrenArray = ChildrenProp->ContainerPtrToValuePtr<TArray<AFGHologram*>>(ParentHologram);
-        if (!ChildrenArray)
-        {
-            return;
-        }
-
-        int32 RemovedCount = 0;
-        for (int32 Index = ChildrenArray->Num() - 1; Index >= 0; --Index)
-        {
-            if (!(*ChildrenArray)[Index] || !IsValid((*ChildrenArray)[Index]))
-            {
-                ChildrenArray->RemoveAt(Index);
-                RemovedCount++;
-            }
-        }
-
-        if (RemovedCount > 0)
-        {
-            SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Warning,
-                TEXT("[SmartRestore][Extend] Scrubbed %d null/invalid parent children: context=%s parent=%s"),
-                RemovedCount,
-                Context ? Context : TEXT("Unknown"),
-                *GetNameSafe(ParentHologram));
-        }
-    }
-}
-
 USFExtendService::USFExtendService()
 {
 }
@@ -296,7 +99,15 @@ void USFExtendService::Initialize(USFSubsystem* InSubsystem)
     WiringService = NewObject<USFExtendWiringService>(this);
     WiringService->Initialize(InSubsystem, this);
 
-    UE_LOG(LogSmartExtend, Log, TEXT("Smart!: SFExtendService initialized (with DetectionService, TopologyService, HologramService, and WiringService)"));
+    // Create and initialize diagnostics service
+    DiagnosticsService = NewObject<USFExtendDiagnosticsService>(this);
+    DiagnosticsService->Initialize(InSubsystem, this);
+
+    // Create and initialize restore-replay service (Smart Restore Extend replay)
+    RestoreReplayService = NewObject<USFExtendRestoreReplayService>(this);
+    RestoreReplayService->Initialize(this);
+
+    UE_LOG(LogSmartExtend, Log, TEXT("Smart!: SFExtendService initialized (with DetectionService, TopologyService, HologramService, WiringService, DiagnosticsService, and RestoreReplayService)"));
 }
 
 // ==================== Mode Management ====================
@@ -357,6 +168,20 @@ void USFExtendService::ClearExtendState()
 void USFExtendService::Shutdown()
 {
     ClearExtendState();
+
+    // Shutdown restore-replay service
+    if (RestoreReplayService)
+    {
+        RestoreReplayService->Shutdown();
+        RestoreReplayService = nullptr;
+    }
+
+    // Shutdown diagnostics service
+    if (DiagnosticsService)
+    {
+        DiagnosticsService->Shutdown();
+        DiagnosticsService = nullptr;
+    }
 
     // Shutdown wiring service
     if (WiringService)
@@ -660,820 +485,36 @@ TSharedPtr<FSFCloneTopology> USFExtendService::GetLastCloneTopology() const
 
 bool USFExtendService::IsHologramCompatibleWithRestoredCloneTopology(AFGHologram* ParentHologram) const
 {
-    if (!bRestoredCloneTopologyActive || !RestoredCloneTopologyTemplate.IsValid())
-    {
-        return true;
-    }
-
-    if (!IsValid(ParentHologram))
-    {
-        return false;
-    }
-
-    const FString& ExpectedBuildClass = RestoredCloneTopologyTemplate->ParentBuildClass;
-    if (ExpectedBuildClass.IsEmpty())
-    {
-        return true;
-    }
-
-    UClass* ActiveBuildClass = ParentHologram->GetBuildClass();
-    const FString ActiveBuildClassName = ActiveBuildClass ? ActiveBuildClass->GetName() : FString();
-    const bool bCompatible = ActiveBuildClassName == ExpectedBuildClass;
-    if (!bCompatible)
-    {
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Log,
-            TEXT("[SmartRestore][Extend] Restored topology build-class mismatch: expected=%s active=%s parent=%s"),
-            *ExpectedBuildClass,
-            ActiveBuildClassName.IsEmpty() ? TEXT("<none>") : *ActiveBuildClassName,
-            *GetNameSafe(ParentHologram));
-    }
-
-    return bCompatible;
+    return !RestoreReplayService || RestoreReplayService->IsHologramCompatibleWithRestoredCloneTopology(ParentHologram);
 }
 
 void USFExtendService::ClearRestoredCloneTopologySession(const TCHAR* Reason)
 {
-    const int32 TemplateChildCount = RestoredCloneTopologyTemplate.IsValid()
-        ? RestoredCloneTopologyTemplate->ChildHolograms.Num()
-        : 0;
-
-    ClearRestoredCloneTopologyPreview();
-    RestoredCloneParentHologram.Reset();
-    RestoredCloneTopologyTemplate.Reset();
-    RestoredCloneBaseTopology.Reset();
-    LastCloneTopology.Reset();
-    RestoredCloneLastParentLocation = FVector::ZeroVector;
-    RestoredCloneLastParentRotation = FRotator::ZeroRotator;
-    bRestoredCloneTopologyActive = false;
-
-    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Log,
-        TEXT("[SmartRestore][Extend] Cleared restored topology session: reason=%s templateChildren=%d"),
-        Reason ? Reason : TEXT("Unknown"),
-        TemplateChildCount);
+    if (RestoreReplayService)
+    {
+        RestoreReplayService->ClearRestoredCloneTopologySession(Reason);
+    }
 }
 
 bool USFExtendService::ReplayRestoreCloneTopology(AFGHologram* ParentHologram, const FSFCloneTopology& CloneTopology)
 {
-    if (!ParentHologram)
-    {
-        return false;
-    }
-
-    if (!CloneTopology.ParentBuildClass.IsEmpty())
-    {
-        UClass* ActiveBuildClass = ParentHologram->GetBuildClass();
-        const FString ActiveBuildClassName = ActiveBuildClass ? ActiveBuildClass->GetName() : FString();
-        if (ActiveBuildClassName != CloneTopology.ParentBuildClass)
-        {
-            SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Log,
-                TEXT("[SmartRestore][Extend] Replay skipped for mismatched build class: expected=%s active=%s parent=%s"),
-                *CloneTopology.ParentBuildClass,
-                ActiveBuildClassName.IsEmpty() ? TEXT("<none>") : *ActiveBuildClassName,
-                *GetNameSafe(ParentHologram));
-            return false;
-        }
-    }
-
-    ClearRestoredCloneTopologyPreview();
-    if (ScaledExtendClones.Num() > 0)
-    {
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Log,
-            TEXT("[SmartRestore][Extend] Clearing %d normal Scaled Extend clone record(s) before Restore replay owns the topology"),
-            ScaledExtendClones.Num());
-        ScaledExtendClones.Empty();
-    }
-    RestoredCloneParentHologram = ParentHologram;
-    RestoredCloneTopologyTemplate = MakeShared<FSFCloneTopology>(CloneTopology);
-    RestoredCloneBaseTopology = MakeShared<FSFCloneTopology>(CloneTopology);
-    RestoredCloneLastParentLocation = ParentHologram->GetActorLocation();
-    RestoredCloneLastParentRotation = ParentHologram->GetActorRotation();
-    bRestoredCloneTopologyActive = true;
-    if (Subsystem.IsValid())
-    {
-        Subsystem->ClearNormalGridChildrenForExtendSuppression(TEXT("ReplayRestoreCloneTopology"));
-    }
-    KickRestoredPreviewParent(ParentHologram);
-
-    FSFCloneTopology ReplayTopology = BuildRestoredCloneTopologyForCurrentState(ParentHologram);
-    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Log,
-        TEXT("[SmartRestore][Extend] Replay restored scaled topology: parent=%s templateChildren=%d replayChildren=%d"),
-        *GetNameSafe(ParentHologram),
-        CloneTopology.ChildHolograms.Num(),
-        ReplayTopology.ChildHolograms.Num());
-    return SpawnRestoredCloneTopology(ParentHologram, ReplayTopology);
+    return RestoreReplayService && RestoreReplayService->ReplayRestoreCloneTopology(ParentHologram, CloneTopology);
 }
 
 void USFExtendService::TickRestoredCloneTopology(float DeltaTime)
 {
-    if (!bRestoredCloneTopologyActive || !HologramService)
+    if (RestoreReplayService)
     {
-        return;
+        RestoreReplayService->TickRestoredCloneTopology(DeltaTime);
     }
-
-    if (!RestoredCloneParentHologram.IsValid())
-    {
-        const bool bHasPostBuildActorsOrRetry = JsonBuiltActors.Num() > 0
-            || bRestoredScaledWiringDeferred
-            || bRestoredScaledWiringRetryScheduled;
-        const bool bCanStillFinishPostBuildWiring = bHasPostBuildActorsOrRetry
-            && HasPendingPostBuildWiring()
-            && (bRestoredScaledWiringRetryScheduled || RestoredScaledWiringRetryAttempts < 5);
-        if (bCanStillFinishPostBuildWiring)
-        {
-            SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Log,
-                TEXT("[SmartRestore][Extend] Parent hologram invalid while post-build wiring is pending; retaining restored topology (storedChildren=%d, jsonBuilt=%d, jsonSpawned=%d, previewFactories=%d, retry=%d/%d)"),
-                StoredCloneTopology.IsValid() ? StoredCloneTopology->ChildHolograms.Num() : 0,
-                JsonBuiltActors.Num(),
-                JsonSpawnedHolograms.Num(),
-                RestoredScaledFactoryPreviewLocations.Num(),
-                RestoredScaledWiringRetryAttempts,
-                5);
-            return;
-        }
-
-        ClearRestoredCloneTopologySession(TEXT("TickRestoredCloneTopology parent invalid"));
-        if (Subsystem.IsValid())
-        {
-            if (USFRestoreService* RestoreSvc = Subsystem->GetRestoreService())
-            {
-                RestoreSvc->ClearActiveRestoreSession(TEXT("Restored Extend parent invalid"));
-            }
-        }
-        return;
-    }
-
-    AFGHologram* ParentHologram = RestoredCloneParentHologram.Get();
-    if (!IsHologramCompatibleWithRestoredCloneTopology(ParentHologram))
-    {
-        ClearRestoredCloneTopologySession(TEXT("TickRestoredCloneTopology build class mismatch"));
-        if (Subsystem.IsValid())
-        {
-            if (USFRestoreService* RestoreSvc = Subsystem->GetRestoreService())
-            {
-                RestoreSvc->ClearActiveRestoreSession(TEXT("Restored Extend parent build class mismatch"));
-            }
-        }
-        return;
-    }
-
-    const FVector ParentLocation = ParentHologram->GetActorLocation();
-    const FRotator ParentRotation = ParentHologram->GetActorRotation();
-    if (!ParentLocation.Equals(RestoredCloneLastParentLocation, SF_RESTORED_PARENT_TRANSFORM_TOLERANCE) || !ParentRotation.Equals(RestoredCloneLastParentRotation, SF_RESTORED_PARENT_TRANSFORM_TOLERANCE))
-    {
-        KickRestoredPreviewParent(ParentHologram);
-        FSFCloneTopology ReplayTopology = BuildRestoredCloneTopologyForCurrentState(ParentHologram);
-        TMap<FString, FVector> IntendedPositions;
-        TMap<FString, FRotator> IntendedRotations;
-        for (const FSFCloneHologram& Holo : ReplayTopology.ChildHolograms)
-        {
-            IntendedPositions.Add(Holo.HologramId, Holo.Transform.Location.ToFVector());
-            IntendedRotations.Add(Holo.HologramId, Holo.Transform.Rotation.ToFRotator());
-        }
-        if (Subsystem.IsValid())
-        {
-            const FSFCounterState& State = Subsystem->GetCounterState();
-            const int32 XCount = FMath::Max(1, FMath::Abs(State.GridCounters.X));
-            const int32 YCount = FMath::Max(1, FMath::Abs(State.GridCounters.Y));
-            if (XCount > 1 || YCount > 1)
-            {
-                const FSFCloneTopology* TemplateTopology = RestoredCloneTopologyTemplate.IsValid()
-                    ? RestoredCloneTopologyTemplate.Get()
-                    : nullptr;
-                for (int32 Y = 0; Y < YCount; ++Y)
-                {
-                    for (int32 X = 0; X < XCount; ++X)
-                    {
-                        if (X == 0 && Y == 0)
-                        {
-                            continue;
-                        }
-
-                        const FRestoredScaledClonePlacement Placement = CalculateRestoredScaledClonePlacement(ParentHologram, TemplateTopology, State, X, Y);
-                        const FString FactoryId = FString::Printf(TEXT("rr_%d_%d_factory"), X, Y);
-                        IntendedPositions.Add(FactoryId, ParentLocation + Placement.WorldOffset);
-                        IntendedRotations.Add(FactoryId, ParentRotation + Placement.RotationOffset);
-                    }
-                }
-            }
-        }
-
-        for (const auto& Pair : JsonSpawnedHolograms)
-        {
-            AFGHologram* Child = Pair.Value;
-            if (!IsValid(Child))
-            {
-                continue;
-            }
-
-            FVector IntendedPos = Child->GetActorLocation();
-            FRotator IntendedRot = Child->GetActorRotation();
-            if (FVector* FoundPos = IntendedPositions.Find(Pair.Key))
-            {
-                IntendedPos = *FoundPos;
-            }
-            if (FRotator* FoundRot = IntendedRotations.Find(Pair.Key))
-            {
-                IntendedRot = *FoundRot;
-            }
-
-            Child->SetActorLocation(IntendedPos);
-            Child->SetActorRotation(IntendedRot);
-            if (USceneComponent* Root = Child->GetRootComponent())
-            {
-                Root->SetWorldLocation(IntendedPos);
-                Root->SetWorldRotation(IntendedRot);
-                Root->MarkRenderStateDirty();
-            }
-            HologramService->TrackChildHologram(Child, IntendedPos, IntendedRot);
-        }
-
-        StoredCloneTopology = MakeShared<FSFCloneTopology>(ReplayTopology);
-        RestoredCloneLastParentLocation = ParentLocation;
-        RestoredCloneLastParentRotation = ParentRotation;
-    }
-
-    HologramService->RefreshChildPositions();
 }
 
 void USFExtendService::OnRestoredCloneTopologyStateChanged()
 {
-    if (!bRestoredCloneTopologyActive || !RestoredCloneParentHologram.IsValid())
+    if (RestoreReplayService)
     {
-        return;
+        RestoreReplayService->OnRestoredCloneTopologyStateChanged();
     }
-
-    AFGHologram* ParentHologram = RestoredCloneParentHologram.Get();
-    FSFCloneTopology ReplayTopology = BuildRestoredCloneTopologyForCurrentState(ParentHologram);
-    ClearRestoredCloneTopologyPreview();
-    RestoredCloneParentHologram = ParentHologram;
-    bRestoredCloneTopologyActive = true;
-    if (Subsystem.IsValid())
-    {
-        Subsystem->ClearNormalGridChildrenForExtendSuppression(TEXT("OnRestoredCloneTopologyStateChanged"));
-    }
-    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Log,
-        TEXT("[SmartRestore][Extend] Restored scaled state changed: parent=%s replayChildren=%d"),
-        *GetNameSafe(ParentHologram),
-        ReplayTopology.ChildHolograms.Num());
-    SpawnRestoredCloneTopology(ParentHologram, ReplayTopology);
-}
-
-FSFCloneTopology USFExtendService::BuildRestoredCloneTopologyForCurrentState(AFGHologram* ParentHologram) const
-{
-    FSFCloneTopology ReplayTopology = RestoredCloneTopologyTemplate.IsValid()
-        ? *RestoredCloneTopologyTemplate
-        : FSFCloneTopology();
-    const FVector OriginalParentLocation = ReplayTopology.ParentTransform.Location.ToFVector();
-    const FRotator OriginalParentRotation = ReplayTopology.ParentTransform.Rotation.ToFRotator();
-    const FVector NewParentLocation = ParentHologram->GetActorLocation();
-    const FRotator NewParentRotation = ParentHologram->GetActorRotation();
-    const FRotator RotationDelta = NewParentRotation - OriginalParentRotation;
-
-    auto TransformLocation = [&](const FVector& Location) -> FVector
-    {
-        const FVector Relative = Location - OriginalParentLocation;
-        return NewParentLocation + RotationDelta.RotateVector(Relative);
-    };
-
-    auto TransformRotation = [&](const FRotator& Rotation) -> FRotator
-    {
-        return Rotation + RotationDelta;
-    };
-
-    ReplayTopology.ParentTransform = FSFTransform(NewParentLocation, NewParentRotation);
-
-    for (FSFCloneHologram& Holo : ReplayTopology.ChildHolograms)
-    {
-        Holo.Transform = FSFTransform(
-            TransformLocation(Holo.Transform.Location.ToFVector()),
-            TransformRotation(Holo.Transform.Rotation.ToFRotator()));
-
-        if (Holo.bHasSplineData)
-        {
-            for (FSFSplinePoint& Point : Holo.SplineData.Points)
-            {
-                Point.World = FSFVec3(TransformLocation(Point.World.ToFVector()));
-            }
-        }
-
-        if (Holo.bHasLiftData)
-        {
-            // TopTransform is actor-local; only the world-space bottom transform follows the parent.
-            Holo.LiftData.BottomTransform = FSFTransform(
-                TransformLocation(Holo.LiftData.BottomTransform.Location.ToFVector()),
-                TransformRotation(Holo.LiftData.BottomTransform.Rotation.ToFRotator()));
-        }
-
-        if (Holo.bIsLaneSegment)
-        {
-            Holo.LaneStartNormal = FSFVec3(RotationDelta.RotateVector(Holo.LaneStartNormal.ToFVector()));
-            Holo.LaneEndNormal = FSFVec3(RotationDelta.RotateVector(Holo.LaneEndNormal.ToFVector()));
-        }
-    }
-
-    const TArray<FSFCloneHologram> SourceChildHolograms = ReplayTopology.ChildHolograms;
-    ReplayTopology.ChildHolograms.RemoveAll([](const FSFCloneHologram& Holo)
-    {
-        return Holo.bIsLaneSegment
-            || Holo.HologramId.StartsWith(TEXT("wire_source_"))
-            || Holo.CloneConnections.ConveyorAny0.Target.StartsWith(TEXT("source:"))
-            || Holo.CloneConnections.ConveyorAny1.Target.StartsWith(TEXT("source:"));
-    });
-
-    if (Subsystem.IsValid() && SourceChildHolograms.Num() > 0)
-    {
-        const FSFCounterState& State = Subsystem->GetCounterState();
-        const int32 XCount = FMath::Max(1, FMath::Abs(State.GridCounters.X));
-        const int32 YCount = FMath::Max(1, FMath::Abs(State.GridCounters.Y));
-        if (XCount > 1 || YCount > 1)
-        {
-            TArray<FSFCloneHologram> ExpandedChildHolograms = ReplayTopology.ChildHolograms;
-            const FSFCloneTopology* TemplateTopology = RestoredCloneTopologyTemplate.IsValid()
-                ? RestoredCloneTopologyTemplate.Get()
-                : nullptr;
-            FVector RestoredSourceFactoryLocation = ParentHologram->GetActorLocation();
-            if (TemplateTopology)
-            {
-                const FRotator TemplateParentRotation = TemplateTopology->ParentTransform.Rotation.ToFRotator();
-                const FRotator TemplateRotationDelta = ParentHologram->GetActorRotation() - TemplateParentRotation;
-                RestoredSourceFactoryLocation -= TemplateRotationDelta.RotateVector(TemplateTopology->WorldOffset.ToFVector());
-            }
-            auto PrefixInternalTarget = [](const FString& InPrefix, const FString& Target) -> FString
-            {
-                if (!Target.IsEmpty()
-                    && Target != TEXT("parent")
-                    && Target != TEXT("external")
-                    && !Target.StartsWith(TEXT("source:")))
-                {
-                    return InPrefix + Target;
-                }
-                return Target;
-            };
-            for (int32 Y = 0; Y < YCount; ++Y)
-            {
-                for (int32 X = 0; X < XCount; ++X)
-                {
-                    if (X == 0 && Y == 0)
-                    {
-                        continue;
-                    }
-
-                    const FString Prefix = FString::Printf(TEXT("rr_%d_%d_"), X, Y);
-                    const FString FactoryId = Prefix + TEXT("factory");
-                    const FString PreviousPrefix = ((X - 1) == 0 && Y == 0)
-                        ? FString()
-                        : FString::Printf(TEXT("rr_%d_%d_"), X - 1, Y);
-                    auto ResolveTargetForCurrentClone = [&](const FString& Target) -> FString
-                    {
-                        if (Target == TEXT("parent"))
-                        {
-                            return FactoryId;
-                        }
-                        return PrefixInternalTarget(Prefix, Target);
-                    };
-                    const FRestoredScaledClonePlacement Placement = CalculateRestoredScaledClonePlacement(ParentHologram, TemplateTopology, State, X, Y);
-                    const FRestoredScaledClonePlacement PreviousPlacement = (X - 1 == 0 && Y == 0)
-                        ? FRestoredScaledClonePlacement()
-                        : CalculateRestoredScaledClonePlacement(ParentHologram, TemplateTopology, State, X - 1, Y);
-                    const FVector ParentLocation = ParentHologram->GetActorLocation();
-                    const FVector CurrentFactoryCenter = ParentLocation + Placement.WorldOffset;
-                    const FVector PreviousFactoryCenter = ParentLocation + PreviousPlacement.WorldOffset;
-
-                    for (FSFCloneHologram Holo : SourceChildHolograms)
-                    {
-                        if (Holo.HologramId.StartsWith(TEXT("wire_source_")))
-                        {
-                            continue;
-                        }
-                        if (Holo.bIsLaneSegment && X == 0)
-                        {
-                            continue;
-                        }
-
-                        const FString OriginalC0Target = Holo.CloneConnections.ConveyorAny0.Target;
-                        const FString OriginalC1Target = Holo.CloneConnections.ConveyorAny1.Target;
-                        if (!Holo.bIsLaneSegment
-                            && (OriginalC0Target.StartsWith(TEXT("source:")) || OriginalC1Target.StartsWith(TEXT("source:"))))
-                        {
-                            continue;
-                        }
-                        Holo.HologramId = Prefix + Holo.HologramId;
-                        if (Holo.bIsLaneSegment && OriginalC0Target.StartsWith(TEXT("source:")))
-                        {
-                            Holo.CloneConnections.ConveyorAny0.Target = PrefixInternalTarget(PreviousPrefix, OriginalC1Target);
-                            Holo.CloneConnections.ConveyorAny1.Target = PrefixInternalTarget(Prefix, OriginalC1Target);
-                            Holo.LaneFromDistributorId = Holo.CloneConnections.ConveyorAny0.Target;
-                            Holo.LaneToDistributorId = Holo.CloneConnections.ConveyorAny1.Target;
-                        }
-                        else if (Holo.bIsLaneSegment && OriginalC1Target.StartsWith(TEXT("source:")))
-                        {
-                            Holo.CloneConnections.ConveyorAny0.Target = PrefixInternalTarget(Prefix, OriginalC0Target);
-                            Holo.CloneConnections.ConveyorAny1.Target = PrefixInternalTarget(PreviousPrefix, OriginalC0Target);
-                            Holo.LaneFromDistributorId = Holo.CloneConnections.ConveyorAny0.Target;
-                            Holo.LaneToDistributorId = Holo.CloneConnections.ConveyorAny1.Target;
-                        }
-                        else
-                        {
-                            Holo.CloneConnections.ConveyorAny0.Target = ResolveTargetForCurrentClone(OriginalC0Target);
-                            Holo.CloneConnections.ConveyorAny1.Target = ResolveTargetForCurrentClone(OriginalC1Target);
-                        }
-                        Holo.ConnectedPowerPoleHologramId = PrefixInternalTarget(Prefix, Holo.ConnectedPowerPoleHologramId);
-                        if (Holo.bHasLiftData)
-                        {
-                            for (FString& PassthroughCloneId : Holo.LiftData.PassthroughCloneIds)
-                            {
-                                PassthroughCloneId = PrefixInternalTarget(Prefix, PassthroughCloneId);
-                            }
-                        }
-                        if (Holo.bIsLaneSegment && Holo.bHasSplineData && Holo.SplineData.Points.Num() >= 2)
-                        {
-                            const bool bSourceAtStart = OriginalC0Target.StartsWith(TEXT("source:"));
-                            const bool bSourceAtEnd = OriginalC1Target.StartsWith(TEXT("source:"));
-                            const FVector BaseStart = Holo.SplineData.Points[0].World.ToFVector();
-                            const FVector BaseEnd = Holo.SplineData.Points.Last().World.ToFVector();
-
-                            auto MoveSourceEndpointToPreviousClone = [&](const FVector& SourceEndpoint) -> FVector
-                            {
-                                const FVector RelativeToSourceFactory = SourceEndpoint - RestoredSourceFactoryLocation;
-                                return PreviousFactoryCenter + PreviousPlacement.RotationOffset.RotateVector(RelativeToSourceFactory);
-                            };
-                            auto MoveCloneEndpointToCurrentClone = [&](const FVector& CloneEndpoint) -> FVector
-                            {
-                                const FVector RelativeToParentFactory = CloneEndpoint - ParentLocation;
-                                return CurrentFactoryCenter + Placement.RotationOffset.RotateVector(RelativeToParentFactory);
-                            };
-
-                            FVector NewStart = BaseStart;
-                            FVector NewEnd = BaseEnd;
-                            if (bSourceAtStart)
-                            {
-                                NewStart = MoveSourceEndpointToPreviousClone(BaseStart);
-                                NewEnd = MoveCloneEndpointToCurrentClone(BaseEnd);
-                                Holo.LaneStartNormal = FSFVec3(PreviousPlacement.RotationOffset.RotateVector(Holo.LaneStartNormal.ToFVector()));
-                                Holo.LaneEndNormal = FSFVec3(Placement.RotationOffset.RotateVector(Holo.LaneEndNormal.ToFVector()));
-                            }
-                            else if (bSourceAtEnd)
-                            {
-                                NewStart = MoveCloneEndpointToCurrentClone(BaseStart);
-                                NewEnd = MoveSourceEndpointToPreviousClone(BaseEnd);
-                                Holo.LaneStartNormal = FSFVec3(Placement.RotationOffset.RotateVector(Holo.LaneStartNormal.ToFVector()));
-                                Holo.LaneEndNormal = FSFVec3(PreviousPlacement.RotationOffset.RotateVector(Holo.LaneEndNormal.ToFVector()));
-                            }
-                            else
-                            {
-                                NewStart = MoveCloneEndpointToCurrentClone(BaseStart);
-                                NewEnd = MoveCloneEndpointToCurrentClone(BaseEnd);
-                                Holo.LaneStartNormal = FSFVec3(Placement.RotationOffset.RotateVector(Holo.LaneStartNormal.ToFVector()));
-                                Holo.LaneEndNormal = FSFVec3(Placement.RotationOffset.RotateVector(Holo.LaneEndNormal.ToFVector()));
-                            }
-
-                            const uint8 FlippedLaneNormalMask = AlignRestoredLaneNormals(Holo, NewStart, NewEnd);
-                            const float NewLength = FVector::Dist(NewStart, NewEnd);
-                            Holo.Transform = FSFTransform(NewStart, (NewEnd - NewStart).Rotation());
-                            Holo.SplineData.Length = NewLength;
-                            Holo.SplineData.Points[0].World = FSFVec3(NewStart);
-                            Holo.SplineData.Points[0].Local = FSFVec3(FVector::ZeroVector);
-                            Holo.SplineData.Points.Last().World = FSFVec3(NewEnd);
-                            Holo.SplineData.Points.Last().Local = FSFVec3(FVector(NewLength, 0.0f, 0.0f));
-
-                            SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Log,
-                                TEXT("[SmartRestore][Extend] Restored lane chained: id=%s grid=(%d,%d) sourceSide=%s prev=%s current=%s length=%.0f"),
-                                *Holo.HologramId,
-                                X,
-                                Y,
-                                bSourceAtStart ? TEXT("start") : (bSourceAtEnd ? TEXT("end") : TEXT("unknown")),
-                                *PreviousPrefix,
-                                *Prefix,
-                                NewLength);
-                            if (FlippedLaneNormalMask != 0)
-                            {
-                                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Log,
-                                    TEXT("[SmartRestore][Extend] Restored lane normals aligned: id=%s type=%s grid=(%d,%d) flipMask=%d"),
-                                    *Holo.HologramId,
-                                    *Holo.LaneSegmentType,
-                                    X,
-                                    Y,
-                                    static_cast<int32>(FlippedLaneNormalMask));
-                            }
-                        }
-                        else
-                        {
-                            const FVector RelativeHoloLocation = Holo.Transform.Location.ToFVector() - ParentLocation;
-                            Holo.Transform.Location = FSFVec3(CurrentFactoryCenter + Placement.RotationOffset.RotateVector(RelativeHoloLocation));
-                            Holo.Transform.Rotation = FSFRot3(Holo.Transform.Rotation.ToFRotator() + Placement.RotationOffset);
-                        }
-                        if (!Holo.bIsLaneSegment && Holo.bHasSplineData)
-                        {
-                            for (FSFSplinePoint& Point : Holo.SplineData.Points)
-                            {
-                                const FVector RelativePointLocation = Point.World.ToFVector() - ParentLocation;
-                                Point.World = FSFVec3(CurrentFactoryCenter + Placement.RotationOffset.RotateVector(RelativePointLocation));
-                            }
-                        }
-                        if (!Holo.bIsLaneSegment && Holo.bHasLiftData)
-                        {
-                            const FVector RelativeBottomLocation = Holo.LiftData.BottomTransform.Location.ToFVector() - ParentLocation;
-                            Holo.LiftData.BottomTransform.Location = FSFVec3(CurrentFactoryCenter + Placement.RotationOffset.RotateVector(RelativeBottomLocation));
-                            Holo.LiftData.BottomTransform.Rotation = FSFRot3(Holo.LiftData.BottomTransform.Rotation.ToFRotator() + Placement.RotationOffset);
-                        }
-                        if (Holo.bIsLaneSegment && (!Holo.bHasSplineData || Holo.SplineData.Points.Num() < 2))
-                        {
-                            Holo.LaneStartNormal = FSFVec3(Placement.RotationOffset.RotateVector(Holo.LaneStartNormal.ToFVector()));
-                            Holo.LaneEndNormal = FSFVec3(Placement.RotationOffset.RotateVector(Holo.LaneEndNormal.ToFVector()));
-                        }
-                        ExpandedChildHolograms.Add(Holo);
-                    }
-                }
-            }
-            ReplayTopology.ChildHolograms = ExpandedChildHolograms;
-            SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Log,
-                TEXT("[SmartRestore][Extend] Restored scaled topology expanded: grid=(%d,%d) children=%d baseChildren=%d"),
-                State.GridCounters.X,
-                State.GridCounters.Y,
-                ReplayTopology.ChildHolograms.Num(),
-                ExpandedChildHolograms.Num() / FMath::Max(1, XCount * YCount));
-        }
-    }
-
-    return ReplayTopology;
-}
-
-void USFExtendService::ClearRestoredCloneTopologyPreview()
-{
-    if (HologramService)
-    {
-        if (RestoredCloneParentHologram.IsValid())
-        {
-            HologramService->SetCurrentParentHologram(RestoredCloneParentHologram.Get());
-        }
-        HologramService->ClearBeltPreviews();
-    }
-    JsonSpawnedHolograms.Empty();
-    RestoredScaledFactoryPreviewLocations.Empty();
-    StoredCloneTopology.Reset();
-    bRestoredCloneTopologyActive = false;
-    bRestoredScaledWiringDeferred = false;
-    bRestoredScaledWiringRetryScheduled = false;
-    RestoredScaledWiringRetryAttempts = 0;
-}
-
-int32 USFExtendService::SpawnRestoredScaledFactoryHolograms(AFGHologram* ParentHologram, TMap<FString, AFGHologram*>& OutSpawnedHolograms)
-{
-    if (!ParentHologram || !Subsystem.IsValid())
-    {
-        return 0;
-    }
-
-    if (FSFHologramData* ParentData = USFHologramDataService::GetOrCreateData(ParentHologram))
-    {
-        ParentData->JsonCloneId = TEXT("parent");
-    }
-
-    const FSFCounterState& State = Subsystem->GetCounterState();
-    const int32 XCount = FMath::Max(1, FMath::Abs(State.GridCounters.X));
-    const int32 YCount = FMath::Max(1, FMath::Abs(State.GridCounters.Y));
-    RestoredScaledFactoryPreviewLocations.Empty();
-    if (XCount <= 1 && YCount <= 1)
-    {
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Log,
-            TEXT("[SmartRestore][Extend] Restored scaled factories: base parent only parent=%s"),
-            *GetNameSafe(ParentHologram));
-        return 0;
-    }
-
-    const FVector ParentLocation = ParentHologram->GetActorLocation();
-    const FRotator ParentRotation = ParentHologram->GetActorRotation();
-    const FSFCloneTopology* TemplateTopology = RestoredCloneTopologyTemplate.IsValid()
-        ? RestoredCloneTopologyTemplate.Get()
-        : nullptr;
-    int32 SpawnedFactories = 0;
-
-    for (int32 Y = 0; Y < YCount; ++Y)
-    {
-        for (int32 X = 0; X < XCount; ++X)
-        {
-            if (X == 0 && Y == 0)
-            {
-                continue;
-            }
-
-            const FRestoredScaledClonePlacement Placement = CalculateRestoredScaledClonePlacement(ParentHologram, TemplateTopology, State, X, Y);
-            const FVector FactoryLocation = ParentLocation + Placement.WorldOffset;
-            const FString FactoryId = FString::Printf(TEXT("rr_%d_%d_factory"), X, Y);
-            static int32 RestoredScaledFactoryCounter = 0;
-            const FName ChildName(*FString::Printf(TEXT("RestoredFactory_%d_%d_%d"), X, Y, RestoredScaledFactoryCounter++));
-
-            AFGHologram* FactoryHologram = AFGHologram::SpawnChildHologramFromRecipe(
-                ParentHologram,
-                ChildName,
-                ParentHologram->GetRecipe(),
-                ParentHologram->GetOwner() ? ParentHologram->GetOwner() : ParentHologram,
-                FactoryLocation,
-                nullptr);
-
-            if (!FactoryHologram)
-            {
-                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Warning,
-                    TEXT("[SmartRestore][Extend] Restored scaled factory spawn failed: id=%s grid=(%d,%d) parent=%s"),
-                    *FactoryId,
-                    X,
-                    Y,
-                    *GetNameSafe(ParentHologram));
-                continue;
-            }
-
-            FactoryHologram->SetActorLocation(FactoryLocation);
-            FactoryHologram->SetActorRotation(ParentRotation + Placement.RotationOffset);
-            FactoryHologram->SetActorHiddenInGame(false);
-            if (USceneComponent* Root = FactoryHologram->GetRootComponent())
-            {
-                Root->SetWorldLocation(FactoryLocation);
-                Root->SetWorldRotation(ParentRotation + Placement.RotationOffset);
-                Root->MarkRenderStateDirty();
-            }
-            FactoryHologram->UpdateComponentTransforms();
-            FactoryHologram->Tags.AddUnique(FName(TEXT("SF_ExtendChild")));
-            USFHologramDataService::DisableValidation(FactoryHologram);
-            USFHologramDataService::MarkAsChild(FactoryHologram, ParentHologram, ESFChildHologramType::ExtendClone);
-
-            if (USFRecipeManagementService* RecipeSvc = Subsystem->GetRecipeManagementService())
-            {
-                if (RecipeSvc->HasStoredProductionRecipe() && RecipeSvc->GetStoredProductionRecipe())
-                {
-                    USFHologramDataService::StoreRecipe(FactoryHologram, RecipeSvc->GetStoredProductionRecipe());
-                }
-            }
-
-            if (FSFHologramData* FactoryData = USFHologramDataService::GetOrCreateData(FactoryHologram))
-            {
-                FactoryData->JsonCloneId = FactoryId;
-            }
-
-            TArray<UBoxComponent*> BoxComponents;
-            FactoryHologram->GetComponents<UBoxComponent>(BoxComponents);
-            for (UBoxComponent* Box : BoxComponents)
-            {
-                Box->SetCollisionEnabled(ECollisionEnabled::NoCollision);
-                Box->SetGenerateOverlapEvents(false);
-            }
-
-            FactoryHologram->SetPlacementMaterialState(EHologramMaterialState::HMS_OK);
-            FactoryHologram->SetActorTickEnabled(false);
-            OutSpawnedHolograms.Add(FactoryId, FactoryHologram);
-            RestoredScaledFactoryPreviewLocations.Add(FactoryId, FactoryLocation);
-            SpawnedFactories++;
-
-            SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Log,
-                TEXT("[SmartRestore][Extend] Restored scaled factory spawned: id=%s grid=(%d,%d) loc=%s"),
-                *FactoryId,
-                X,
-                Y,
-                *FactoryLocation.ToString());
-        }
-    }
-
-    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Log,
-        TEXT("[SmartRestore][Extend] Restored scaled factories spawned: count=%d grid=(%d,%d) parent=%s"),
-        SpawnedFactories,
-        State.GridCounters.X,
-        State.GridCounters.Y,
-        *GetNameSafe(ParentHologram));
-    return SpawnedFactories;
-}
-
-bool USFExtendService::SpawnRestoredCloneTopology(AFGHologram* ParentHologram, const FSFCloneTopology& CloneTopology)
-{
-    TMap<FString, AFGHologram*> SpawnedHolograms;
-    const int32 SpawnedFactoryCount = SpawnRestoredScaledFactoryHolograms(ParentHologram, SpawnedHolograms);
-    const int32 SpawnedCount = CloneTopology.SpawnChildHolograms(ParentHologram, this, SpawnedHolograms);
-    ScrubInvalidHologramChildren(ParentHologram, TEXT("SpawnRestoredCloneTopology"));
-    if (SpawnedCount <= 0)
-    {
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Warning,
-            TEXT("[SmartRestore][Extend] ReplayRestoreCloneTopology failed: spawned=%d factories=%d inputChildren=%d parent=%s"),
-            SpawnedCount,
-            SpawnedFactoryCount,
-            CloneTopology.ChildHolograms.Num(),
-            *GetNameSafe(ParentHologram));
-        return false;
-    }
-
-    CloneTopology.WireChildHologramConnections(SpawnedHolograms, ParentHologram);
-    StoredCloneTopology = MakeShared<FSFCloneTopology>(CloneTopology);
-    JsonSpawnedHolograms = SpawnedHolograms;
-    if (HologramService)
-    {
-        HologramService->SetCurrentParentHologram(ParentHologram);
-        HologramService->StoreCloneTopology(StoredCloneTopology);
-        HologramService->StoreJsonSpawnedHolograms(SpawnedHolograms);
-        TMap<FString, FVector> IntendedPositions;
-        TMap<FString, FRotator> IntendedRotations;
-        TMap<FString, const FSFCloneHologram*> HologramDataMap;
-        for (const FSFCloneHologram& Holo : CloneTopology.ChildHolograms)
-        {
-            IntendedPositions.Add(Holo.HologramId, Holo.Transform.Location.ToFVector());
-            IntendedRotations.Add(Holo.HologramId, Holo.Transform.Rotation.ToFRotator());
-            HologramDataMap.Add(Holo.HologramId, &Holo);
-        }
-        for (const auto& Pair : SpawnedHolograms)
-        {
-            if (AFGHologram* Child = Pair.Value)
-            {
-                FVector IntendedPos = Child->GetActorLocation();
-                FRotator IntendedRot = Child->GetActorRotation();
-                if (FVector* FoundPos = IntendedPositions.Find(Pair.Key))
-                {
-                    IntendedPos = *FoundPos;
-                }
-                if (FRotator* FoundRot = IntendedRotations.Find(Pair.Key))
-                {
-                    IntendedRot = *FoundRot;
-                }
-                Child->SetActorLocation(IntendedPos);
-                Child->SetActorRotation(IntendedRot);
-                if (USceneComponent* Root = Child->GetRootComponent())
-                {
-                    Root->SetWorldLocation(IntendedPos);
-                    Root->SetWorldRotation(IntendedRot);
-                }
-                Child->UpdateComponentTransforms();
-                if (const FSFCloneHologram** HoloDataPtr = HologramDataMap.Find(Pair.Key))
-                {
-                    const FSFCloneHologram& HoloData = **HoloDataPtr;
-                    if (ASFConveyorBeltHologram* Belt = Cast<ASFConveyorBeltHologram>(Child))
-                    {
-                        if (HoloData.bIsLaneSegment && HoloData.bHasSplineData && HoloData.SplineData.Points.Num() >= 2)
-                        {
-                            const FVector StartPos = HoloData.SplineData.Points[0].World.ToFVector();
-                            const FVector EndPos = HoloData.SplineData.Points.Last().World.ToFVector();
-                            const FVector StartNormal = HoloData.LaneStartNormal.ToFVector();
-                            const FVector EndNormal = HoloData.LaneEndNormal.ToFVector();
-                            Belt->AutoRouteSplineWithNormals(StartPos, StartNormal, EndPos, EndNormal);
-                            Belt->TriggerMeshGeneration();
-                            Belt->ForceApplyHologramMaterial();
-                        }
-                        else if (HoloData.bHasSplineData)
-                        {
-                            TArray<FSplinePointData> SplinePoints;
-                            for (const FSFSplinePoint& Point : HoloData.SplineData.Points)
-                            {
-                                FSplinePointData PointData;
-                                PointData.Location = Point.Local.ToFVector();
-                                PointData.ArriveTangent = Point.ArriveTangent.ToFVector();
-                                PointData.LeaveTangent = Point.LeaveTangent.ToFVector();
-                                SplinePoints.Add(PointData);
-                            }
-                            Belt->SetSplineDataAndUpdate(SplinePoints);
-                            Belt->TriggerMeshGeneration();
-                            Belt->ForceApplyHologramMaterial();
-                        }
-                    }
-                    else if (ASFPipelineHologram* Pipe = Cast<ASFPipelineHologram>(Child))
-                    {
-                        if (HoloData.bIsLaneSegment && HoloData.bHasSplineData && HoloData.SplineData.Points.Num() >= 2)
-                        {
-                            const FVector StartPos = HoloData.SplineData.Points[0].World.ToFVector();
-                            const FVector EndPos = HoloData.SplineData.Points.Last().World.ToFVector();
-                            const FVector StartNormal = HoloData.LaneStartNormal.ToFVector();
-                            const FVector EndNormal = HoloData.LaneEndNormal.ToFVector();
-                            Pipe->TryUseBuildModeRouting(StartPos, StartNormal, EndPos, EndNormal);
-                            Pipe->TriggerMeshGeneration();
-                            Pipe->ForceApplyHologramMaterial();
-                        }
-                        else if (HoloData.bHasSplineData)
-                        {
-                            TArray<FSplinePointData> SplinePoints;
-                            for (const FSFSplinePoint& Point : HoloData.SplineData.Points)
-                            {
-                                FSplinePointData PointData;
-                                PointData.Location = Point.Local.ToFVector();
-                                PointData.ArriveTangent = Point.ArriveTangent.ToFVector();
-                                PointData.LeaveTangent = Point.LeaveTangent.ToFVector();
-                                SplinePoints.Add(PointData);
-                            }
-                            Pipe->SetSplineDataAndUpdate(SplinePoints);
-                            Pipe->TriggerMeshGeneration();
-                            Pipe->ForceApplyHologramMaterial();
-                        }
-                    }
-                }
-                HologramService->TrackChildHologram(Child, IntendedPos, IntendedRot);
-            }
-        }
-        KickRestoredPreviewParent(ParentHologram);
-        HologramService->RefreshChildPositions();
-    }
-
-    RestoredCloneLastParentLocation = ParentHologram->GetActorLocation();
-    RestoredCloneLastParentRotation = ParentHologram->GetActorRotation();
-
-    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Log,
-        TEXT("[SmartRestore][Extend] Replayed Extend topology: infra=%d factories=%d tracked=%d parent=%s"),
-        SpawnedCount,
-        SpawnedFactoryCount,
-        HologramService ? HologramService->GetTrackedChildren().Num() : 0,
-        *GetNameSafe(ParentHologram));
-    return true;
 }
 
 void USFExtendService::ClearTopology()
@@ -2502,7 +1543,7 @@ void USFExtendService::CreateBeltPreviews(AFGHologram* ParentHologram)
     }
 
     // DIAGNOSTIC: Capture preview snapshot ONCE when EXTEND previews are first created
-    if (!bHasPreviewSnapshot)
+    if (!HasPreviewSnapshot())
     {
         CapturePreviewSnapshot();
     }
@@ -5610,865 +4651,27 @@ AFGBuildableFactory* USFExtendService::GetSourceFactory() const
     return nullptr;
 }
 
-// ==================== Diagnostic Capture ====================
-
-FSFBuildableSnapshot USFExtendService::CaptureNearbyBuildables(float Radius)
-{
-    FSFBuildableSnapshot Snapshot;
-    Snapshot.CaptureRadius = Radius;
-    Snapshot.CaptureTime = FDateTime::Now();
-
-    if (!Subsystem.IsValid())
-    {
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Warning, TEXT("📊 DIAG: Cannot capture - no subsystem"));
-        return Snapshot;
-    }
-
-    // Get player location
-    APlayerController* PC = Subsystem->GetWorld()->GetFirstPlayerController();
-    if (!PC || !PC->GetPawn())
-    {
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Warning, TEXT("📊 DIAG: Cannot capture - no player"));
-        return Snapshot;
-    }
-
-    FVector PlayerLocation = PC->GetPawn()->GetActorLocation();
-
-    // ==================== Build EXTEND Topology Lookup ====================
-    // Maps buildable actor -> (Role, ChainId, ChainIndex)
-    struct FExtendTopologyInfo {
-        FString Role;
-        int32 ChainId = -1;
-        int32 ChainIndex = -1;
-    };
-    TMap<AActor*, FExtendTopologyInfo> TopologyLookup;
-
-    // Add source factory
-    if (CurrentExtendTarget.IsValid())
-    {
-        FExtendTopologyInfo Info;
-        Info.Role = TEXT("SourceFactory");
-        TopologyLookup.Add(CurrentExtendTarget.Get(), Info);
-    }
-
-    // Add input belt chain members (conveyors = belts + lifts, poles, and distributors)
-    for (int32 ChainIdx = 0; ChainIdx < GetCurrentTopology().InputChains.Num(); ++ChainIdx)
-    {
-        const FSFConnectionChainNode& Chain = GetCurrentTopology().InputChains[ChainIdx];
-
-        // Add distributor (splitter) at chain end
-        if (Chain.Distributor.IsValid())
-        {
-            FExtendTopologyInfo Info;
-            Info.Role = TEXT("Splitter");
-            Info.ChainId = ChainIdx;
-            TopologyLookup.Add(Chain.Distributor.Get(), Info);
-        }
-
-        // Add all conveyors (belts and lifts) in chain
-        for (int32 ConvIdx = 0; ConvIdx < Chain.Conveyors.Num(); ++ConvIdx)
-        {
-            if (Chain.Conveyors[ConvIdx].IsValid())
-            {
-                AFGBuildableConveyorBase* Conv = Chain.Conveyors[ConvIdx].Get();
-                FExtendTopologyInfo Info;
-                // Distinguish belt from lift
-                Info.Role = Cast<AFGBuildableConveyorLift>(Conv) ? TEXT("InputLift") : TEXT("InputBelt");
-                Info.ChainId = ChainIdx;
-                Info.ChainIndex = ConvIdx;
-                TopologyLookup.Add(Conv, Info);
-            }
-        }
-
-        // Add all support poles in chain (future cloning candidate)
-        for (int32 PoleIdx = 0; PoleIdx < Chain.SupportPoles.Num(); ++PoleIdx)
-        {
-            if (Chain.SupportPoles[PoleIdx].IsValid())
-            {
-                FExtendTopologyInfo Info;
-                Info.Role = TEXT("Pole");
-                Info.ChainId = ChainIdx;
-                Info.ChainIndex = PoleIdx;
-                TopologyLookup.Add(Chain.SupportPoles[PoleIdx].Get(), Info);
-            }
-        }
-    }
-
-    // Add output belt chain members
-    for (int32 ChainIdx = 0; ChainIdx < GetCurrentTopology().OutputChains.Num(); ++ChainIdx)
-    {
-        const FSFConnectionChainNode& Chain = GetCurrentTopology().OutputChains[ChainIdx];
-
-        // Add distributor (merger) at chain end
-        if (Chain.Distributor.IsValid())
-        {
-            FExtendTopologyInfo Info;
-            Info.Role = TEXT("Merger");
-            Info.ChainId = ChainIdx;
-            TopologyLookup.Add(Chain.Distributor.Get(), Info);
-        }
-
-        // Add all conveyors (belts and lifts) in chain
-        for (int32 ConvIdx = 0; ConvIdx < Chain.Conveyors.Num(); ++ConvIdx)
-        {
-            if (Chain.Conveyors[ConvIdx].IsValid())
-            {
-                AFGBuildableConveyorBase* Conv = Chain.Conveyors[ConvIdx].Get();
-                FExtendTopologyInfo Info;
-                Info.Role = Cast<AFGBuildableConveyorLift>(Conv) ? TEXT("OutputLift") : TEXT("OutputBelt");
-                Info.ChainId = ChainIdx;
-                Info.ChainIndex = ConvIdx;
-                TopologyLookup.Add(Conv, Info);
-            }
-        }
-
-        // Add all support poles in chain
-        for (int32 PoleIdx = 0; PoleIdx < Chain.SupportPoles.Num(); ++PoleIdx)
-        {
-            if (Chain.SupportPoles[PoleIdx].IsValid())
-            {
-                FExtendTopologyInfo Info;
-                Info.Role = TEXT("Pole");
-                Info.ChainId = ChainIdx;
-                Info.ChainIndex = PoleIdx;
-                TopologyLookup.Add(Chain.SupportPoles[PoleIdx].Get(), Info);
-            }
-        }
-    }
-
-    // Add input pipe chain members
-    for (int32 ChainIdx = 0; ChainIdx < GetCurrentTopology().PipeInputChains.Num(); ++ChainIdx)
-    {
-        const FSFPipeConnectionChainNode& Chain = GetCurrentTopology().PipeInputChains[ChainIdx];
-
-        // Add junction at chain end
-        if (Chain.Junction.IsValid())
-        {
-            FExtendTopologyInfo Info;
-            Info.Role = TEXT("Junction");
-            Info.ChainId = ChainIdx;
-            TopologyLookup.Add(Chain.Junction.Get(), Info);
-        }
-
-        // Add all pipes in chain
-        for (int32 PipeIdx = 0; PipeIdx < Chain.Pipelines.Num(); ++PipeIdx)
-        {
-            if (Chain.Pipelines[PipeIdx].IsValid())
-            {
-                FExtendTopologyInfo Info;
-                Info.Role = TEXT("InputPipe");
-                Info.ChainId = ChainIdx;
-                Info.ChainIndex = PipeIdx;
-                TopologyLookup.Add(Chain.Pipelines[PipeIdx].Get(), Info);
-            }
-        }
-
-        // Add all support poles (pipe supports) in chain
-        for (int32 PoleIdx = 0; PoleIdx < Chain.SupportPoles.Num(); ++PoleIdx)
-        {
-            if (Chain.SupportPoles[PoleIdx].IsValid())
-            {
-                FExtendTopologyInfo Info;
-                Info.Role = TEXT("PipeSupport");
-                Info.ChainId = ChainIdx;
-                Info.ChainIndex = PoleIdx;
-                TopologyLookup.Add(Chain.SupportPoles[PoleIdx].Get(), Info);
-            }
-        }
-    }
-
-    // Add output pipe chain members
-    for (int32 ChainIdx = 0; ChainIdx < GetCurrentTopology().PipeOutputChains.Num(); ++ChainIdx)
-    {
-        const FSFPipeConnectionChainNode& Chain = GetCurrentTopology().PipeOutputChains[ChainIdx];
-
-        // Add junction at chain end
-        if (Chain.Junction.IsValid())
-        {
-            FExtendTopologyInfo Info;
-            Info.Role = TEXT("Junction");
-            Info.ChainId = ChainIdx;
-            TopologyLookup.Add(Chain.Junction.Get(), Info);
-        }
-
-        // Add all pipes in chain
-        for (int32 PipeIdx = 0; PipeIdx < Chain.Pipelines.Num(); ++PipeIdx)
-        {
-            if (Chain.Pipelines[PipeIdx].IsValid())
-            {
-                FExtendTopologyInfo Info;
-                Info.Role = TEXT("OutputPipe");
-                Info.ChainId = ChainIdx;
-                Info.ChainIndex = PipeIdx;
-                TopologyLookup.Add(Chain.Pipelines[PipeIdx].Get(), Info);
-            }
-        }
-
-        // Add all support poles (pipe supports) in chain
-        for (int32 PoleIdx = 0; PoleIdx < Chain.SupportPoles.Num(); ++PoleIdx)
-        {
-            if (Chain.SupportPoles[PoleIdx].IsValid())
-            {
-                FExtendTopologyInfo Info;
-                Info.Role = TEXT("PipeSupport");
-                Info.ChainId = ChainIdx;
-                Info.ChainIndex = PoleIdx;
-                TopologyLookup.Add(Chain.SupportPoles[PoleIdx].Get(), Info);
-            }
-        }
-    }
-
-    UE_LOG(LogSmartExtend, VeryVerbose, TEXT("📊 DIAG: Built topology lookup with %d members"), TopologyLookup.Num());
-    Snapshot.CaptureLocation = PlayerLocation;
-
-    // Find all buildables in radius
-    TArray<AActor*> FoundActors;
-    UGameplayStatics::GetAllActorsOfClass(Subsystem->GetWorld(), AFGBuildable::StaticClass(), FoundActors);
-
-    for (AActor* Actor : FoundActors)
-    {
-        AFGBuildable* Buildable = Cast<AFGBuildable>(Actor);
-        if (!Buildable) continue;
-
-        float Distance = FVector::Dist(PlayerLocation, Buildable->GetActorLocation());
-        if (Distance > Radius) continue;
-
-        FSFCapturedBuildable Captured;
-
-        // === Identity ===
-        Captured.Name = Buildable->GetName();
-        Captured.ClassName = Buildable->GetClass()->GetName();
-
-        // === Transform ===
-        Captured.Location = Buildable->GetActorLocation();
-        Captured.Rotation = Buildable->GetActorRotation();
-        Captured.Scale = Buildable->GetActorScale3D();
-
-        // Get bounds
-        FVector Origin, BoxExtent;
-        Buildable->GetActorBounds(false, Origin, BoxExtent);
-        Captured.BoundsMin = Origin - BoxExtent;
-        Captured.BoundsMax = Origin + BoxExtent;
-
-        // === State ===
-        Captured.bIsHidden = Buildable->IsHidden();
-        Captured.bIsPendingKill = !IsValid(Buildable);
-        Captured.bHasBegunPlay = Buildable->HasActorBegunPlay();
-
-        // === Capture Factory Connections ===
-        TArray<UFGFactoryConnectionComponent*> FactoryConns;
-        Buildable->GetComponents<UFGFactoryConnectionComponent>(FactoryConns);
-        for (UFGFactoryConnectionComponent* Conn : FactoryConns)
-        {
-            if (!Conn) continue;
-
-            FSFCapturedConnection CapturedConn;
-            CapturedConn.ConnectorName = Conn->GetName();
-            CapturedConn.ConnectorClass = Conn->GetClass()->GetName();
-            CapturedConn.WorldLocation = Conn->GetComponentLocation();
-            CapturedConn.WorldRotation = Conn->GetComponentRotation();
-            CapturedConn.Direction = (int32)Conn->GetDirection();
-            CapturedConn.bIsConnected = Conn->IsConnected();
-
-            if (Conn->IsConnected())
-            {
-                UFGFactoryConnectionComponent* OtherConn = Conn->GetConnection();
-                if (OtherConn && OtherConn->GetOwner())
-                {
-                    CapturedConn.ConnectedToActor = OtherConn->GetOwner()->GetName();
-                    CapturedConn.ConnectedToConnector = OtherConn->GetName();
-                }
-            }
-
-            Captured.FactoryConnections.Add(CapturedConn);
-        }
-
-        // === Capture Pipe Connections ===
-        TArray<UFGPipeConnectionComponent*> PipeConns;
-        Buildable->GetComponents<UFGPipeConnectionComponent>(PipeConns);
-        for (UFGPipeConnectionComponent* Conn : PipeConns)
-        {
-            if (!Conn) continue;
-
-            FSFCapturedConnection CapturedConn;
-            CapturedConn.ConnectorName = Conn->GetName();
-            CapturedConn.ConnectorClass = Conn->GetClass()->GetName();
-            CapturedConn.WorldLocation = Conn->GetComponentLocation();
-            CapturedConn.WorldRotation = Conn->GetComponentRotation();
-            CapturedConn.Direction = (int32)Conn->GetPipeConnectionType();
-            CapturedConn.bIsConnected = Conn->IsConnected();
-
-            if (Conn->IsConnected())
-            {
-                UFGPipeConnectionComponentBase* OtherConnBase = Conn->GetConnection();
-                if (OtherConnBase && OtherConnBase->GetOwner())
-                {
-                    CapturedConn.ConnectedToActor = OtherConnBase->GetOwner()->GetName();
-                    CapturedConn.ConnectedToConnector = OtherConnBase->GetName();
-                }
-            }
-
-            Captured.PipeConnections.Add(CapturedConn);
-        }
-
-        // === Categorize and gather type-specific data ===
-        if (AFGBuildableConveyorBelt* Belt = Cast<AFGBuildableConveyorBelt>(Buildable))
-        {
-            Captured.Category = TEXT("Belt");
-            Captured.BeltSpeed = Belt->GetSpeed();
-
-            // Get spline data
-            if (USplineComponent* Spline = Belt->GetSplineComponent())
-            {
-                Captured.SplineLength = Spline->GetSplineLength();
-                Captured.SplinePointCount = Spline->GetNumberOfSplinePoints();
-
-                for (int32 i = 0; i < Spline->GetNumberOfSplinePoints(); ++i)
-                {
-                    FSFCapturedSplinePoint Point;
-                    Point.Index = i;
-                    Point.Location = Spline->GetLocationAtSplinePoint(i, ESplineCoordinateSpace::Local);
-                    Point.WorldLocation = Spline->GetLocationAtSplinePoint(i, ESplineCoordinateSpace::World);
-                    Point.ArriveTangent = Spline->GetArriveTangentAtSplinePoint(i, ESplineCoordinateSpace::World);
-                    Point.LeaveTangent = Spline->GetLeaveTangentAtSplinePoint(i, ESplineCoordinateSpace::World);
-                    Point.Rotation = Spline->GetRotationAtSplinePoint(i, ESplineCoordinateSpace::World);
-                    Captured.SplinePoints.Add(Point);
-                }
-            }
-        }
-        else if (AFGBuildableConveyorLift* Lift = Cast<AFGBuildableConveyorLift>(Buildable))
-        {
-            Captured.Category = TEXT("Lift");
-            Captured.LiftHeight = Lift->GetHeight();
-            Captured.bLiftIsReversed = Lift->GetIsReversed();
-
-            // Get connection locations
-            TArray<UFGFactoryConnectionComponent*> LiftConns;
-            Lift->GetComponents<UFGFactoryConnectionComponent>(LiftConns);
-            for (UFGFactoryConnectionComponent* Conn : LiftConns)
-            {
-                if (Conn->GetName().Contains(TEXT("0")))
-                {
-                    Captured.LiftBottomLocation = Conn->GetComponentLocation();
-                }
-                else
-                {
-                    Captured.LiftTopLocation = Conn->GetComponentLocation();
-                }
-            }
-        }
-        else if (Cast<AFGBuildableConveyorAttachment>(Buildable))
-        {
-            Captured.Category = TEXT("Distributor");
-        }
-        else if (AFGBuildablePipeline* Pipe = Cast<AFGBuildablePipeline>(Buildable))
-        {
-            Captured.Category = TEXT("Pipe");
-
-            // Get spline data
-            if (USplineComponent* Spline = Pipe->GetSplineComponent())
-            {
-                Captured.SplineLength = Spline->GetSplineLength();
-                Captured.SplinePointCount = Spline->GetNumberOfSplinePoints();
-
-                for (int32 i = 0; i < Spline->GetNumberOfSplinePoints(); ++i)
-                {
-                    FSFCapturedSplinePoint Point;
-                    Point.Index = i;
-                    Point.Location = Spline->GetLocationAtSplinePoint(i, ESplineCoordinateSpace::Local);
-                    Point.WorldLocation = Spline->GetLocationAtSplinePoint(i, ESplineCoordinateSpace::World);
-                    Point.ArriveTangent = Spline->GetArriveTangentAtSplinePoint(i, ESplineCoordinateSpace::World);
-                    Point.LeaveTangent = Spline->GetLeaveTangentAtSplinePoint(i, ESplineCoordinateSpace::World);
-                    Point.Rotation = Spline->GetRotationAtSplinePoint(i, ESplineCoordinateSpace::World);
-                    Captured.SplinePoints.Add(Point);
-                }
-            }
-        }
-        else if (Cast<AFGBuildablePipelineJunction>(Buildable))
-        {
-            Captured.Category = TEXT("Junction");
-        }
-        else if (Cast<AFGBuildableFactory>(Buildable))
-        {
-            Captured.Category = TEXT("Factory");
-        }
-        else
-        {
-            Captured.Category = TEXT("Other");
-        }
-
-        // === Check if this buildable is part of EXTEND topology ===
-        if (FExtendTopologyInfo* TopoInfo = TopologyLookup.Find(Buildable))
-        {
-            Captured.bIsExtendSource = true;
-            Captured.ExtendRole = TopoInfo->Role;
-            Captured.ExtendChainId = TopoInfo->ChainId;
-            Captured.ExtendChainIndex = TopoInfo->ChainIndex;
-        }
-
-        Snapshot.Buildables.Add(Captured);
-
-        // Update category count
-        int32& Count = Snapshot.CountByCategory.FindOrAdd(Captured.Category);
-        Count++;
-    }
-
-    return Snapshot;
-}
+// ==================== Diagnostic Capture (delegates to DiagnosticsService) ====================
 
 void USFExtendService::CapturePreviewSnapshot()
 {
-    // DISABLED: RadarPulse generates too much log output (~1400 lines per Extend)
-    // Re-enable for debugging by uncommenting the block below
-    /*
-    // Use RadarPulse service for enhanced capture (200m radius)
-    USFRadarPulseService* RadarPulse = Subsystem.IsValid() ? Subsystem->GetRadarPulseService() : nullptr;
-    if (RadarPulse)
+    if (DiagnosticsService)
     {
-        // Capture with RadarPulse (200m = 20000cm)
-        FSFRadarPulseSnapshot PulseSnapshot = RadarPulse->CaptureSnapshot(20000.0f, TEXT("EXTEND_PRE"));
-
-        // Flag EXTEND source objects
-        RadarPulse->FlagExtendSourceObjects(PulseSnapshot, GetCurrentTopology(), CurrentExtendTarget.Get());
-
-        // Cache for later comparison
-        RadarPulse->CacheSnapshot(TEXT("EXTEND_PRE"), PulseSnapshot);
-
-        // Log the snapshot with EXTEND source details
-        // RadarPulse->LogFlaggedObjects(PulseSnapshot, TEXT("ExtendSource"), true);  // DISABLED: Too verbose (~2000 lines)
-
-        bHasPreviewSnapshot = true;
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Log, TEXT("📡 RadarPulse: EXTEND preview snapshot captured (%d objects, %d EXTEND sources)"),
-            PulseSnapshot.TotalObjects, PulseSnapshot.ExtendSourceCount);
-        return;
-    }
-    */
-
-    // Fallback to legacy capture if RadarPulse unavailable
-    PreviewSnapshot = CaptureNearbyBuildables(15000.0f);  // 150m
-    bHasPreviewSnapshot = true;
-
-    UE_LOG(LogSmartExtend, VeryVerbose, TEXT("📊 DIAG: Captured PREVIEW snapshot - %d buildables within 150m"),
-        PreviewSnapshot.Buildables.Num());
-
-    // Log summary by category
-    for (const auto& Pair : PreviewSnapshot.CountByCategory)
-    {
-        UE_LOG(LogSmartExtend, VeryVerbose, TEXT("📊 DIAG:   %s: %d"), *Pair.Key, Pair.Value);
-    }
-
-    // === Log EXTEND Source Topology Summary ===
-    int32 SourceCount = 0;
-    TMap<FString, int32> SourceByRole;
-    for (const FSFCapturedBuildable& B : PreviewSnapshot.Buildables)
-    {
-        if (B.bIsExtendSource)
-        {
-            SourceCount++;
-            int32& Count = SourceByRole.FindOrAdd(B.ExtendRole);
-            Count++;
-        }
-    }
-
-    if (SourceCount > 0)
-    {
-        // DISABLED: Source topology box logging - generates ~30+ lines per Extend
-        // Re-enable for debugging by uncommenting the block below
-        /*
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT(""));
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("╔═══════════════════════════════════════════════════════════════════╗"));
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("║           EXTEND SOURCE TOPOLOGY (%d members)                      ║"), SourceCount);
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("╠═══════════════════════════════════════════════════════════════════╣"));
-
-        for (const auto& Pair : SourceByRole)
-        {
-            SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("║   %-20s: %3d                                       ║"), *Pair.Key, Pair.Value);
-        }
-
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("╚═══════════════════════════════════════════════════════════════════╝"));
-        */
-
-        // DISABLED: Detailed source topology logging generates ~1500 lines per Extend
-        // Re-enable for debugging by uncommenting the block below
-        /*
-        // Log full details of SOURCE topology members
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT(""));
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("📊 EXTEND SOURCE DETAILS:"));
-
-        for (const FSFCapturedBuildable& B : PreviewSnapshot.Buildables)
-        {
-            if (B.bIsExtendSource)
-            {
-                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT(""));
-                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("┌─────────────────────────────────────────────────────────────────────"));
-                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ ★ %s [%s] - Chain %d, Index %d"),
-                    *B.ExtendRole, *B.Name, B.ExtendChainId, B.ExtendChainIndex);
-                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("├─────────────────────────────────────────────────────────────────────"));
-                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ Category: %s | Class: %s"), *B.Category, *B.ClassName);
-                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ Location: X=%.3f Y=%.3f Z=%.3f"), B.Location.X, B.Location.Y, B.Location.Z);
-                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ Rotation: P=%.3f Y=%.3f R=%.3f"), B.Rotation.Pitch, B.Rotation.Yaw, B.Rotation.Roll);
-
-                // Belt-specific data
-                if (B.Category == TEXT("Belt"))
-                {
-                    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ ═══ BELT DATA ═══"));
-                    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ Speed: %.1f | SplineLength: %.1fcm | SplinePoints: %d"),
-                        B.BeltSpeed, B.SplineLength, B.SplinePointCount);
-
-                    for (const FSFCapturedSplinePoint& SP : B.SplinePoints)
-                    {
-                        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│   [Point %d] World=(%.1f,%.1f,%.1f)"),
-                            SP.Index, SP.WorldLocation.X, SP.WorldLocation.Y, SP.WorldLocation.Z);
-                        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│             Arrive=(%.1f,%.1f,%.1f) Leave=(%.1f,%.1f,%.1f)"),
-                            SP.ArriveTangent.X, SP.ArriveTangent.Y, SP.ArriveTangent.Z,
-                            SP.LeaveTangent.X, SP.LeaveTangent.Y, SP.LeaveTangent.Z);
-                    }
-                }
-
-                // Lift-specific data
-                if (B.Category == TEXT("Lift"))
-                {
-                    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ ═══ LIFT DATA ═══"));
-                    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ Height: %.1fcm | Reversed: %d"), B.LiftHeight, B.bLiftIsReversed);
-                    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ Bottom: (%.1f,%.1f,%.1f) Top: (%.1f,%.1f,%.1f)"),
-                        B.LiftBottomLocation.X, B.LiftBottomLocation.Y, B.LiftBottomLocation.Z,
-                        B.LiftTopLocation.X, B.LiftTopLocation.Y, B.LiftTopLocation.Z);
-                }
-
-                // Pipe-specific data
-                if (B.Category == TEXT("Pipe"))
-                {
-                    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ ═══ PIPE DATA ═══"));
-                    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ SplineLength: %.1fcm | SplinePoints: %d"), B.SplineLength, B.SplinePointCount);
-
-                    for (const FSFCapturedSplinePoint& SP : B.SplinePoints)
-                    {
-                        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│   [Point %d] World=(%.1f,%.1f,%.1f)"),
-                            SP.Index, SP.WorldLocation.X, SP.WorldLocation.Y, SP.WorldLocation.Z);
-                    }
-                }
-
-                // Factory connections
-                if (B.FactoryConnections.Num() > 0)
-                {
-                    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ ═══ FACTORY CONNECTIONS (%d) ═══"), B.FactoryConnections.Num());
-                    for (const FSFCapturedConnection& Conn : B.FactoryConnections)
-                    {
-                        FString ConnStatus = Conn.bIsConnected
-                            ? FString::Printf(TEXT("-> %s.%s"), *Conn.ConnectedToActor, *Conn.ConnectedToConnector)
-                            : TEXT("(not connected)");
-                        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│   %s [Dir=%d] @ (%.1f,%.1f,%.1f) %s"),
-                            *Conn.ConnectorName, Conn.Direction,
-                            Conn.WorldLocation.X, Conn.WorldLocation.Y, Conn.WorldLocation.Z,
-                            *ConnStatus);
-                    }
-                }
-
-                // Pipe connections
-                if (B.PipeConnections.Num() > 0)
-                {
-                    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ ═══ PIPE CONNECTIONS (%d) ═══"), B.PipeConnections.Num());
-                    for (const FSFCapturedConnection& Conn : B.PipeConnections)
-                    {
-                        FString ConnStatus = Conn.bIsConnected
-                            ? FString::Printf(TEXT("-> %s.%s"), *Conn.ConnectedToActor, *Conn.ConnectedToConnector)
-                            : TEXT("(not connected)");
-                        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│   %s [Type=%d] @ (%.1f,%.1f,%.1f) %s"),
-                            *Conn.ConnectorName, Conn.Direction,
-                            Conn.WorldLocation.X, Conn.WorldLocation.Y, Conn.WorldLocation.Z,
-                            *ConnStatus);
-                    }
-                }
-
-                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("└─────────────────────────────────────────────────────────────────────"));
-            }
-        }
-        */
+        DiagnosticsService->CapturePreviewSnapshot();
     }
 }
 
 void USFExtendService::CapturePostBuildSnapshotAndLogDiff()
 {
-    if (!bHasPreviewSnapshot)
+    if (DiagnosticsService)
     {
-        UE_LOG(LogSmartExtend, VeryVerbose, TEXT("📊 DIAG: No preview snapshot to compare against"));
-        return;
+        DiagnosticsService->CapturePostBuildSnapshotAndLogDiff();
     }
-
-    // DISABLED: RadarPulse generates too much log output (~1400 lines per Extend)
-    // Re-enable for debugging by uncommenting the block below
-    /*
-    // Use RadarPulse service for enhanced capture and diff
-    USFRadarPulseService* RadarPulse = Subsystem.IsValid() ? Subsystem->GetRadarPulseService() : nullptr;
-    if (RadarPulse)
-    {
-        // Capture POST snapshot (200m = 20000cm)
-        FSFRadarPulseSnapshot PostSnapshot = RadarPulse->CaptureSnapshot(20000.0f, TEXT("EXTEND_POST"));
-
-        // Get the PRE snapshot from cache
-        FSFRadarPulseSnapshot PreSnapshot;
-        if (RadarPulse->GetCachedSnapshot(TEXT("EXTEND_PRE"), PreSnapshot))
-        {
-            // Compare and log the diff
-            FSFSnapshotDiff Diff = RadarPulse->CompareSnapshots(PreSnapshot, PostSnapshot);
-            RadarPulse->LogDiff(Diff, true);  // Verbose output
-
-            SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Log, TEXT("📡 RadarPulse: EXTEND build diff - %d new, %d removed, %d modified"),
-                Diff.NewObjects.Num(), Diff.RemovedObjects.Num(), Diff.ModifiedObjects.Num());
-        }
-        else
-        {
-            // No PRE snapshot cached, just log the POST snapshot
-            RadarPulse->LogSnapshot(PostSnapshot, true);
-        }
-
-        // Clear cache
-        RadarPulse->ClearCache(TEXT("EXTEND_PRE"));
-        bHasPreviewSnapshot = false;
-        return;
-    }
-    */
-
-    // Fallback to legacy capture if RadarPulse unavailable
-    FSFBuildableSnapshot PostBuildSnapshot = CaptureNearbyBuildables(15000.0f);  // 150m
-
-    UE_LOG(LogSmartExtend, VeryVerbose, TEXT("📊 DIAG: Captured POST-BUILD snapshot - %d buildables within 150m"),
-        PostBuildSnapshot.Buildables.Num());
-
-    LogSnapshotDiff(PreviewSnapshot, PostBuildSnapshot);
-
-    // Clear the preview snapshot
-    bHasPreviewSnapshot = false;
-    PreviewSnapshot = FSFBuildableSnapshot();
 }
 
-void USFExtendService::LogSnapshotDiff(const FSFBuildableSnapshot& Before, const FSFBuildableSnapshot& After)
+bool USFExtendService::HasPreviewSnapshot() const
 {
-    // DISABLED: BUILD DIFF SUMMARY box logging - generates ~20+ lines per Extend
-    // Re-enable for debugging by uncommenting the block below
-    /*
-    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT(""));
-    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("╔═══════════════════════════════════════════════════════════════════╗"));
-    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("║               EXTEND DIAGNOSTIC: BUILD DIFF SUMMARY               ║"));
-    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("╠═══════════════════════════════════════════════════════════════════╣"));
-    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("║ Capture Radius: 150m | Before: %4d buildables | After: %4d        ║"),
-        Before.Buildables.Num(), After.Buildables.Num());
-    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("╠═══════════════════════════════════════════════════════════════════╣"));
-
-    // Build set of existing names for quick lookup
-    TSet<FString> BeforeNames;
-    for (const FSFCapturedBuildable& B : Before.Buildables)
-    {
-        BeforeNames.Add(B.Name);
-    }
-
-    TSet<FString> AfterNames;
-    for (const FSFCapturedBuildable& A : After.Buildables)
-    {
-        AfterNames.Add(A.Name);
-    }
-
-    // Find new buildables (in After but not in Before)
-    TArray<FSFCapturedBuildable> NewBuildables;
-    for (const FSFCapturedBuildable& A : After.Buildables)
-    {
-        if (!BeforeNames.Contains(A.Name))
-        {
-            NewBuildables.Add(A);
-        }
-    }
-
-    // Find removed buildables (in Before but not in After)
-    TArray<FSFCapturedBuildable> RemovedBuildables;
-    for (const FSFCapturedBuildable& B : Before.Buildables)
-    {
-        if (!AfterNames.Contains(B.Name))
-        {
-            RemovedBuildables.Add(B);
-        }
-    }
-
-    // Count new by category
-    TMap<FString, int32> NewByCategory;
-    for (const FSFCapturedBuildable& N : NewBuildables)
-    {
-        int32& Count = NewByCategory.FindOrAdd(N.Category);
-        Count++;
-    }
-
-    // Count removed by category
-    TMap<FString, int32> RemovedByCategory;
-    for (const FSFCapturedBuildable& R : RemovedBuildables)
-    {
-        int32& Count = RemovedByCategory.FindOrAdd(R.Category);
-        Count++;
-    }
-
-    // Log category summary
-    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("║ Category        │ Before │ After  │ New    │ Removed │ Delta   ║"));
-    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("╟─────────────────┼────────┼────────┼────────┼─────────┼─────────╢"));
-
-    // Collect all categories
-    TSet<FString> AllCategories;
-    for (const auto& Pair : Before.CountByCategory) AllCategories.Add(Pair.Key);
-    for (const auto& Pair : After.CountByCategory) AllCategories.Add(Pair.Key);
-
-    for (const FString& Category : AllCategories)
-    {
-        int32 BeforeCount = Before.CountByCategory.Contains(Category) ? Before.CountByCategory[Category] : 0;
-        int32 AfterCount = After.CountByCategory.Contains(Category) ? After.CountByCategory[Category] : 0;
-        int32 NewCount = NewByCategory.Contains(Category) ? NewByCategory[Category] : 0;
-        int32 RemovedCount = RemovedByCategory.Contains(Category) ? RemovedByCategory[Category] : 0;
-        int32 Delta = AfterCount - BeforeCount;
-
-        FString DeltaStr = Delta > 0 ? FString::Printf(TEXT("+%d"), Delta) : FString::Printf(TEXT("%d"), Delta);
-
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("║ %-15s │ %6d │ %6d │ %6d │ %7d │ %7s ║"),
-            *Category, BeforeCount, AfterCount, NewCount, RemovedCount, *DeltaStr);
-    }
-
-    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("╠═══════════════════════════════════════════════════════════════════╣"));
-    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("║ TOTAL           │ %6d │ %6d │ %6d │ %7d │ %+7d ║"),
-        Before.Buildables.Num(), After.Buildables.Num(),
-        NewBuildables.Num(), RemovedBuildables.Num(),
-        After.Buildables.Num() - Before.Buildables.Num());
-    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("╚═══════════════════════════════════════════════════════════════════╝"));
-    */
-
-    // DISABLED: Detailed buildable enumeration generates ~1000 lines per Extend
-    // Re-enable for debugging by uncommenting the block below
-    /*
-    // Log FULL details of ALL new buildables
-    if (NewBuildables.Num() > 0)
-    {
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT(""));
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("╔═══════════════════════════════════════════════════════════════════╗"));
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("║            FULL ENUMERATION: NEW BUILDABLES (%d total)            ║"), NewBuildables.Num());
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("╚═══════════════════════════════════════════════════════════════════╝"));
-
-        for (int32 i = 0; i < NewBuildables.Num(); ++i)
-        {
-            const FSFCapturedBuildable& N = NewBuildables[i];
-
-            SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT(""));
-            SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("┌─────────────────────────────────────────────────────────────────────"));
-
-            // Show EXTEND source badge prominently
-            if (N.bIsExtendSource)
-            {
-                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ [%d] ★ EXTEND SOURCE ★ %s"), i, *N.Name);
-                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│     Role: %s | Chain: %d | Index: %d"),
-                    *N.ExtendRole, N.ExtendChainId, N.ExtendChainIndex);
-            }
-            else
-            {
-                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ [%d] %s"), i, *N.Name);
-            }
-
-            SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("├─────────────────────────────────────────────────────────────────────"));
-            SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ Category: %s | Class: %s"), *N.Category, *N.ClassName);
-            SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ Location: X=%.3f Y=%.3f Z=%.3f"), N.Location.X, N.Location.Y, N.Location.Z);
-            SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ Rotation: P=%.3f Y=%.3f R=%.3f"), N.Rotation.Pitch, N.Rotation.Yaw, N.Rotation.Roll);
-            SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ Scale: X=%.3f Y=%.3f Z=%.3f"), N.Scale.X, N.Scale.Y, N.Scale.Z);
-            SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ Bounds: Min=(%.1f,%.1f,%.1f) Max=(%.1f,%.1f,%.1f)"),
-                N.BoundsMin.X, N.BoundsMin.Y, N.BoundsMin.Z, N.BoundsMax.X, N.BoundsMax.Y, N.BoundsMax.Z);
-            SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ State: Hidden=%d PendingKill=%d BegunPlay=%d"),
-                N.bIsHidden, N.bIsPendingKill, N.bHasBegunPlay);
-
-            // Belt-specific data
-            if (N.Category == TEXT("Belt"))
-            {
-                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ ═══ BELT DATA ═══"));
-                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ Speed: %.1f | SplineLength: %.1fcm | SplinePoints: %d"),
-                    N.BeltSpeed, N.SplineLength, N.SplinePointCount);
-
-                for (const FSFCapturedSplinePoint& SP : N.SplinePoints)
-                {
-                    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│   [Point %d] Local=(%.1f,%.1f,%.1f) World=(%.1f,%.1f,%.1f)"),
-                        SP.Index, SP.Location.X, SP.Location.Y, SP.Location.Z,
-                        SP.WorldLocation.X, SP.WorldLocation.Y, SP.WorldLocation.Z);
-                    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│             Arrive=(%.1f,%.1f,%.1f) Leave=(%.1f,%.1f,%.1f)"),
-                        SP.ArriveTangent.X, SP.ArriveTangent.Y, SP.ArriveTangent.Z,
-                        SP.LeaveTangent.X, SP.LeaveTangent.Y, SP.LeaveTangent.Z);
-                }
-            }
-
-            // Lift-specific data
-            if (N.Category == TEXT("Lift"))
-            {
-                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ ═══ LIFT DATA ═══"));
-                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ Height: %.1fcm | Reversed: %d"), N.LiftHeight, N.bLiftIsReversed);
-                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ Bottom: (%.1f,%.1f,%.1f) Top: (%.1f,%.1f,%.1f)"),
-                    N.LiftBottomLocation.X, N.LiftBottomLocation.Y, N.LiftBottomLocation.Z,
-                    N.LiftTopLocation.X, N.LiftTopLocation.Y, N.LiftTopLocation.Z);
-            }
-
-            // Pipe-specific data
-            if (N.Category == TEXT("Pipe"))
-            {
-                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ ═══ PIPE DATA ═══"));
-                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ SplineLength: %.1fcm | SplinePoints: %d"), N.SplineLength, N.SplinePointCount);
-
-                for (const FSFCapturedSplinePoint& SP : N.SplinePoints)
-                {
-                    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│   [Point %d] Local=(%.1f,%.1f,%.1f) World=(%.1f,%.1f,%.1f)"),
-                        SP.Index, SP.Location.X, SP.Location.Y, SP.Location.Z,
-                        SP.WorldLocation.X, SP.WorldLocation.Y, SP.WorldLocation.Z);
-                }
-            }
-
-            // Factory connections
-            if (N.FactoryConnections.Num() > 0)
-            {
-                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ ═══ FACTORY CONNECTIONS (%d) ═══"), N.FactoryConnections.Num());
-                for (const FSFCapturedConnection& Conn : N.FactoryConnections)
-                {
-                    FString ConnStatus = Conn.bIsConnected
-                        ? FString::Printf(TEXT("-> %s.%s"), *Conn.ConnectedToActor, *Conn.ConnectedToConnector)
-                        : TEXT("(not connected)");
-                    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│   %s [Dir=%d] @ (%.1f,%.1f,%.1f) %s"),
-                        *Conn.ConnectorName, Conn.Direction,
-                        Conn.WorldLocation.X, Conn.WorldLocation.Y, Conn.WorldLocation.Z,
-                        *ConnStatus);
-                }
-            }
-
-            // Pipe connections
-            if (N.PipeConnections.Num() > 0)
-            {
-                SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│ ═══ PIPE CONNECTIONS (%d) ═══"), N.PipeConnections.Num());
-                for (const FSFCapturedConnection& Conn : N.PipeConnections)
-                {
-                    FString ConnStatus = Conn.bIsConnected
-                        ? FString::Printf(TEXT("-> %s.%s"), *Conn.ConnectedToActor, *Conn.ConnectedToConnector)
-                        : TEXT("(not connected)");
-                    SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("│   %s [Type=%d] @ (%.1f,%.1f,%.1f) %s"),
-                        *Conn.ConnectorName, Conn.Direction,
-                        Conn.WorldLocation.X, Conn.WorldLocation.Y, Conn.WorldLocation.Z,
-                        *ConnStatus);
-                }
-            }
-
-            SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Display, TEXT("└─────────────────────────────────────────────────────────────────────"));
-        }
-    }
-
-    // Log FULL details of removed buildables
-    if (RemovedBuildables.Num() > 0)
-    {
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Warning, TEXT(""));
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Warning, TEXT("╔═══════════════════════════════════════════════════════════════════╗"));
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Warning, TEXT("║          FULL ENUMERATION: REMOVED BUILDABLES (%d total)          ║"), RemovedBuildables.Num());
-        SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Warning, TEXT("╚═══════════════════════════════════════════════════════════════════╝"));
-
-        for (int32 i = 0; i < RemovedBuildables.Num(); ++i)
-        {
-            const FSFCapturedBuildable& R = RemovedBuildables[i];
-            SF_EXTEND_DIAGNOSTIC_LOG(LogSmartExtend, Warning, TEXT("   [%d] [%s] %s @ (%.0f, %.0f, %.0f)"),
-                i, *R.Category, *R.Name, R.Location.X, R.Location.Y, R.Location.Z);
-        }
-    }
-    */
+    return DiagnosticsService && DiagnosticsService->HasPreviewSnapshot();
 }
 
 // ==================== Phase 5/6: JSON-Based Post-Build Wiring ====================
